services/core-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
celery==5.3.4
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
google-cloud-sql-connector==1.4.3
google-cloud-kms==2.19.1
google-cloud-secret-manager==2.16.4
google-cloud-aiplatform==1.38.1
google-cloud-discoveryengine==0.11.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-resource-manager==1.10.4
google-cloud-aiplatform==1.38.1
google-cloud-discoveryengine==0.11.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
stripe==7.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
httpx==0.25.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-gcp-trace==1.6.0
pytest==7.4.3
pytest-asyncio==0.21.1